Fetch after receiving an input event
Previously we were using the keyup event to decide when to fetch
autocomplete results. This approach resulted in us having to declare a
list of keys to ignore to avoid unecessary / unintended fetches.

Instead of listening for keyup events, we can simply listen for changes to
the text in the input field and fetch based on that.

This approach allows many forms of input that may not trigger a keyup event.
For example right clicking and pasting from clipboard, or tapping a autocorrect
suggestion on a mobile device.